core: Make protocols handle seccfg
Due to 167fe0a1225b2b0ee2c4224ce1ffd5866fa2bb4f introducing the SEJ extensions cmd, we don't need to use the SEJ module anymore, being it unstable and unreliable.
If we ever need to, we can just delegate the CryptoIO to the protocol to implement.

So, now protocols handle it internally, and SecCfgV4 now is just a "builder", with the data with provide.
This allows for a cleaner and more maintainable API, without mixing concerns.